+++ src/components/navigation/NavBar.vue
@@ -24,7 +24,7 @@
 </template>
 
 <script>
-import NavItem from '../navigation/NavItem.vue';
+import NavItem from './NavItem.vue';
 
 export default {
   name: 'nav-bar',
