+++ src/components/views/Post.vue
@@ -47,7 +47,7 @@
 </template>
 
 <script>
-import Divider from '@/components/helpers/Divider.vue';
+import Divider from '@/components/misc/Divider.vue';
 
 export default {
   name: 'post-view',
