+++ src/components/views/PostFeed.vue
@@ -147,7 +147,7 @@
 </template>
 
 <script>
-import Divider from '@/components/helpers/Divider.vue';
+import Divider from '@/components/misc/Divider.vue';
 import PostPreview from '@/components/previews/Post.vue';
 
 export default {
